encrypt_decrypt/Amazon_add1/1/requirements.txt
requests
beautifulsoup4
pyyaml
playwright
